+++ @xen-orchestra/web/src/modules/pool/components/actions/disconnect/PoolDisconnectButton.vue
@@ -1,6 +1,6 @@
 <template>
   <MenuItem
-    accent="brand"
+    accent="neutral"
     icon="action:disconnect"
     class="disconnect typo-body-bold-small"
     :busy="isDisconnectingServer"