+++ @xen-orchestra/web/src/modules/vif/components/actions/delete/VifDeleteButton.vue
@@ -1,5 +1,11 @@
 <template>
-  <MenuItem accent="brand" icon="action:delete" :disabled="!canDeleteVifs" :busy="isDeletingVifs" @click="deleteVifs()">
+  <MenuItem
+    accent="neutral"
+    icon="action:delete"
+    :disabled="!canDeleteVifs"
+    :busy="isDeletingVifs"
+    @click="deleteVifs()"
+  >
     {{ t('action:delete') }}
     <i v-if="hint">{{ hint }}</i>
   </MenuItem>