+++ @xen-orchestra/web/src/modules/vm/components/actions/resume/VmResumeButton.vue
@@ -1,7 +1,7 @@
 <template>
   <MenuItem
     v-if="(isSuspended && canRun) || isRunning"
-    accent="brand"
+    accent="neutral"
     icon="fa:play"
     :busy="isRunning"
     @click="resumeJob()"